requirements.txt
chardet==5.2.0
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
fastrlock==0.8.3
fonttools==4.57.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==1.23.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pandas_ta==0.3.14b0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.0
scikit-learn==1.2.2
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
xgboost==3.0.0
